Integrate match commands into normal mode
