+++ VsHelix/EnterKeyHandler.cs
@@ -37,8 +37,6 @@ namespace VsHelix
 
                         if (ModeManager.Instance.Current == ModeManager.EditorMode.Search && ModeManager.Instance.Search != null)
                                 return ModeManager.Instance.Search.HandleChar('\r', view, broker, ops);
-                        if (ModeManager.Instance.Current == ModeManager.EditorMode.Match && ModeManager.Instance.Match != null)
-                                return ModeManager.Instance.Match.HandleChar('\r', view, broker, ops);
 
                         return false;
                 }
