+++ VsHelix/StatusBarHelper.cs
@@ -14,7 +14,6 @@ namespace VsHelix
 				ModeManager.EditorMode.Normal => "NOR",
 				ModeManager.EditorMode.Insert => "INS",
                                 ModeManager.EditorMode.Search => "SCH",
-                                ModeManager.EditorMode.Match => "MAT",
                                 _ => mode.ToString().ToUpperInvariant(),
 			};
 			if (!string.IsNullOrEmpty(extra))
